Refactor and move things. Figured out how to do data transfer.
