+++ ROVController/src/Core/Event.cpp
@@ -1,0 +1,9 @@
+#include "Event.h"
+
+Core::Event::~Event() {
+	if (type == SensorInfoReceived) {
+		if (sInfo.type == SensorInfo::Video) {
+			delete [] sInfo.f.data;
+		}
+	}
+}
